refactor: move ScrollIntoView to atoms
