+++ src/components/molecules/index.ts
@@ -21,7 +21,6 @@ export {default as ResourceFilterIconWithPopover} from './ResourceFilterIconWith
 export {default as ResourceNamespaceFilter} from './ResourceNamespaceFilter';
 export {default as ResourceRefsIconPopover} from './ResourceRefsIconPopover';
 export {default as SaveResourcesToFileFolderModal} from './SaveResourcesToFileFolderModal';
-export {default as ScrollIntoView} from './ScrollIntoView';
 export {default as SectionRenderer} from './SectionRenderer';
 export {default as TemplateFormRenderer} from './TemplateFormRenderer';
 export {default as TitleBar} from './TitleBar';
